+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.1] - 2023-01-17
+
+- remove default password and replace to random password
+- add custom hook `before_user_save`
+
 ## [0.0.1] - 2023-01-03
 
 - Initial release